+++ rl_2048/__init__.py
@@ -2,8 +2,8 @@
 
 __version__ = "0.0.0"
 
-from rl_2048.DQN.dqn import DQN
-from rl_2048.DQN.net import Net
+from rl_2048.dqn.torch.dqn import DQN
+from rl_2048.dqn.torch.net import Net
 from rl_2048.game_engine import GameEngine
 from rl_2048.tile import Tile
 from rl_2048.tile_plotter import TilePlotter
