+++ tests/jax_dqn/test_replay_memory.py
@@ -1,7 +1,7 @@
 import jax.random as jrandom
 from jax import Array
 
-from rl_2048.jax_dqn.replay_memory import Action, Batch, ReplayMemory, Transition
+from rl_2048.dqn.jax.replay_memory import Action, Batch, ReplayMemory, Transition
 
 all_memory_fields = {"states", "actions", "next_states", "rewards", "games_over"}
